requirements.txt
accelerate==1.11.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.1
aiosignal==1.4.0
altair==5.5.0
annotated-doc==0.0.3
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.11.0
attrs==25.4.0
azure-ai-documentintelligence==1.0.2
azure-common==1.1.28
azure-core==1.36.0
azure-identity==1.25.1
azure-search-documents==11.5.2
beautifulsoup4==4.14.2
blinker==1.9.0
cachetools==6.2.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
colorlog==6.10.1
cryptography==46.0.3
dataclasses-json==0.6.7
dill==0.4.0
distro==1.9.0
docling==2.58.0
docling-core==2.49.0
docling-ibm-models==3.10.1
docling-parse==4.7.0
et_xmlfile==2.0.0
Faker==37.11.0
fastapi==0.120.0
filelock==3.20.0
filetype==1.2.0
frozenlist==1.8.0
fsspec==2025.9.0
gitdb==4.0.12
GitPython==3.1.45
greenlet==3.2.4
h11==0.16.0
hf-xet==1.1.10
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.36.0
idna==3.11
isodate==0.7.2
Jinja2==3.1.6
jiter==0.11.1
jsonlines==4.0.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
langchain==0.3.27
langchain-community==0.3.31
langchain-core==0.3.79
langchain-docling==1.1.0
langchain-openai==0.3.35
langchain-text-splitters==0.3.11
langchainhub==0.1.21
langgraph==1.0.1
langgraph-checkpoint==3.0.0
langgraph-prebuilt==1.0.1
langgraph-sdk==0.2.9
langsmith==0.4.37
latex2mathml==3.78.1
lxml==5.4.0
markdown-it-py==4.0.0
marko==2.2.1
MarkupSafe==3.0.3
marshmallow==3.26.1
mdurl==0.1.2
mpire==2.10.2
mpmath==1.3.0
msal==1.34.0
msal-extensions==1.3.1
multidict==6.7.0
multiprocess==0.70.18
mypy_extensions==1.1.0
narwhals==2.9.0
networkx==3.5
numpy==2.2.6
omegaconf==2.3.0
openai==2.6.0
opencv-python==4.12.0.88
openpyxl==3.1.5
orjson==3.11.3
ormsgpack==1.11.0
packaging==24.2
pandas==2.3.3
pillow==11.3.0
pluggy==1.6.0
polyfactory==2.22.3
propcache==0.4.1
protobuf==6.33.0
psutil==7.1.1
pyarrow==22.0.0
pyclipper==1.3.0.post6
pycparser==2.23
pydantic==2.12.3
pydantic-settings==2.11.0
pydantic_core==2.41.4
pydeck==0.9.1
Pygments==2.19.2
PyJWT==2.10.1
pylatexenc==2.10
pypdfium2==4.30.0
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
python-multipart==0.0.20
python-pptx==1.0.2
pytz==2025.2
pywin32==311
PyYAML==6.0.3
rapidocr==3.4.2
referencing==0.37.0
regex==2025.10.23
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.2.0
rpds-py==0.28.0
rtree==1.4.1
safetensors==0.6.2
scipy==1.16.2
semchunk==2.2.2
shapely==2.1.2
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.44
starlette==0.48.0
streamlit==1.50.0
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.2
tiktoken==0.12.0
tokenizers==0.22.1
toml==0.10.2
torch==2.9.0
torchvision==0.24.0
tornado==6.5.2
tqdm==4.67.1
transformers==4.57.1
typer==0.19.2
types-requests==2.32.4.20250913
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.38.0
watchdog==6.0.0
xlsxwriter==3.2.9
xxhash==3.6.0
yarl==1.22.0
zstandard==0.25.0
